Resolves the swrObjTest_FLAG0_REPAIRING comment collision (both sides
documented the same bit) by keeping master's setter list and this
branch's mechanism note.

Also adopts this branch's canonical names in the swrRace_CalcTargetTurnRate
body that landed in master via #280 -- git merged those files without a
textual conflict, but the renames live in types.h while the uses live in
swrRace.c, so the merge only broke at compile time:

  player->unk1eb8 -> player->remoteState     (swrRace.c:1166)
  player->unk12_1 -> player->wallHitCooldown (swrRace.c:1184)

The stale "tilted/spun" description of 0x264 in the function comment is
corrected to the post-wall-impact window this branch established it to be.

Full dinput.dll build + link clean.

+++ src/General/stdMath.h
@@ -35,7 +35,7 @@
 #define stdMath_ArcSin3_ADDR (0x0048d010)
 
 void stdMath_MultiplyAddClamped(float* res_inout, float value, float multiplier, float min, float max);
-void stdMath_AddScaledValueAndClamp_i32(int* res_inout, int value, float multiplier, int min, int max);
+void stdMath_AddScaledValueAndClamp_i32(int* res_inout, float value, float multiplier, int min, int max);
 
 void stdMath_SinCos(float angle_degrees, float* pSinOut, float* pCosOut);
 float stdMath_Tan(float angle_degrees);
